+++ test/net/req_email_test.dart
@@ -1,39 +1,34 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:formdator/formdator.dart';
 
-/// Most test cases were taken from:
-/// - [valid-email-addresses](https://en.wikipedia.org/wiki/Email_address#Valid_email_addresses)
 void main() {
-  group('Default error messages', () {
-    final reqEmail = ReqEmail();
-    // test('null - blank', () {
-    //   expect(reqEmail(null), 'required email');
-    // });
-    // test('empty - blank', () {
-    //   expect(reqEmail(''), 'required email');
-    // });
-    // test('invalid email', () {
-    //   expect(reqEmail('123@email'), 'malformed email');
-    // });
+  group('ReqEmail', () {
+    const blank = 'required email address';
+    const mal = 'malformed email address';
+    final reqEmail = ReqEmail(blank: blank, mal: mal);
+    final reqEmailLen = ReqEmail.len(30, blank: blank, mal: mal);
+    test('null - blank', () {
+      expect(reqEmail(null), blank);
+      expect(reqEmailLen(null), blank);
+    });
+    test('empty - blank', () {
+      expect(reqEmail(''), blank);
+      expect(reqEmailLen(''), blank);
+    });
+    test('invalid email', () {
+      const invalid = '123@email';
+      expect(reqEmail(invalid), mal);
+      expect(reqEmailLen(invalid), mal);
+    });
     test('valid email', () {
-      expect(reqEmail('a_valid_user@anyemail.com'), null);
+      const valid = 'a_valid_user@anyemail.com';
+      expect(reqEmail(valid), null);
+      expect(reqEmailLen(valid), null);
+    });
+    test('long, valid email', () {
+      const longValid = 'disposable.style.email.with+symbol@example.com';
+      expect(reqEmail(longValid), null);
+      expect(reqEmailLen(longValid), 'too long email');
     });
   });
-  // group('Custom error messages', () {
-  //   const mal = 'malformed email';
-  //   const blank = 'required email';
-  //   final reqEmail = ReqEmail(blank: blank, mal: mal);
-  //   test('null - blank', () {
-  //     expect(reqEmail(null), blank);
-  //   });
-  //   test('empty - blank', () {
-  //     expect(reqEmail(''), blank);
-  //   });
-  //   test('invalid email, no @ character', () {
-  //     expect(reqEmail('Abc.example.com'), mal);
-  //   });
-  //   test('valid email', () {
-  //     expect(reqEmail('user@validemail.com'), null);
-  //   });
-  // });
 }